recoded routes to use controller files which will make the server side code much cleaner
